refactor: flytt valideringsreglar til eigen validering/-mappe
Flyttar dei fire regelgruppene (arbeidOgUttakDeFørsteSeksUkene,
samtidigUttak, farMedmorRundtFødsel, farMedmorMaksToUkerRundtFødsel)
inn i regler/validering/ med eigen index.ts og types.ts.

Fellestypar (i18n, Periode, Feltregel, Feltregelområde) blir verande
i regler/types.ts. regler/index.ts re-eksporterer frå validering/
slik at eksisterande konsumentar ikkje må endrast.
